pyproject.toml
[project]
name = "kg-gen"
version = "0.4.0"
authors = [{ name = "Belinda Mo", email = "justanexperimentpark@gmail.com" }]
description = "Extract a knowledge graph using LLMs from any text or messages array"
readme = "README.md"
requires-python = ">=3.10, <4.0"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
license = "MIT"
dependencies = [
    "nltk",
    "pydantic>=2.0.0",
    "pre-commit>=4.3.0",
    "ty>=0.0.1a21",
    "dspy>=2.6.27",
    "openai>=1.108.1",
    "pyvis>=0.3.2",
    "sentence-transformers>=5.1.0",
    "tqdm>=4.67.1",
    "datasets>=4.1.1",
    "semhash>=0.3.2",
    "inflect>=7.5.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.7.2",
    "rank-bm25>=0.2.2",
    "networkx>=3.0",
    "python-dotenv>=1.0.0",
    "dspy-ai>=2.6.27",
    "pathlib2>=2.3.7",
    "typer>=0.17.4",
    "ruff>=0.14.2",
]

[dependency-groups]
dev = [
    "datasets>=4.1.1",
    "huggingface-hub>=0.35.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=0.21.0",
    "wikipedia-api>=0.8.1",
    "matplotlib>=3.10.7",
]
mcp = [
    "fastmcp>=2.10.6",
    "mcp>=1.12.1",
]

[project.scripts]
kggen = "kg_gen.cli:main"

[project.urls]
Homepage = "https://github.com/stair-lab/kg-gen"
Issues = "https://github.com/stair-lab/kg-gen/issues"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-data]
"kg_gen.utils" = ["template.html"]

[tool.pytest.ini_options]
testpaths = ["tests", "mcp/tests"]
